requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.0
fonttools==4.46.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
Werkzeug==3.0.1
azure-cognitiveservices-speech==1.34.0
azure-storage-blob==12.19.0